Replaced comparison of strings with == with .equals().
As far as I know, comparing strings with == is dangerous, as it does a
reference comparison while .equals() checks for the values. So it is
possible that "test" == "test" evaluates to false, if they are two
different objects, while "test".equals("test") will always return true.

If you *wanted* a reference comparison there, I apologize.